Refactor AbstractShoppingList to AbstractItemList and update usages.
This commit refactors the `AbstractShoppingList` class to `AbstractItemList` for better clarity.
It also includes minor formatting changes in `build.gradle.kts` and `settings.gradle.kts`, and adds new IDEA configuration files for Ktlint and Android test results preferences.
