--- app/src/main/java/com/example/recipeShoppingList/domain/ShoppingList/AbstractShoppingList.kt
@@ -1,18 +1,0 @@
-
-
-abstract class AbstractShoppingList{
-
-    private val shoppingListItems: MutableList<String> = mutableListOf()
-
-    fun getListItems(): List<String> {
-        return shoppingListItems.toList()
-    }
-
-    protected fun addItem(item: String){
-        shoppingListItems.add(item)
-    }
-
-    protected fun removeItem(item: String){
-        shoppingListItems.remove(item)
-    }
-}
